--- domains/certificates/Interfaces/Interfaces.sln
@@ -1,25 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 17
-VisualStudioVersion = 17.3.32811.315
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Interfaces", "Interfaces.csproj", "{7240CE50-FD5C-4F82-9620-DD75290BB674}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{7240CE50-FD5C-4F82-9620-DD75290BB674}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{7240CE50-FD5C-4F82-9620-DD75290BB674}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{7240CE50-FD5C-4F82-9620-DD75290BB674}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{7240CE50-FD5C-4F82-9620-DD75290BB674}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {F95DC181-1E5C-4D2F-95E5-22B3581BF560}
-	EndGlobalSection
-EndGlobal
