+++ domains/certificates/Libraries/CertificateEvents/CertificateRejected.cs
@@ -1,7 +1,7 @@
 using System;
 using EnergyOriginEventStore.EventStore.Serialization;
 
-namespace CertificatesEvents;
+namespace CertificateEvents;
 
 [EventModelVersion("CertificateRejected", 1)]
 public record CertificateRejected(Guid CertificateId, string Reason) : EventModel;
